Repository improvements
ISBC listing
You can now see the list of all the ISBC (IS business case) as you add them.
For now you can only see a brief preview, I'm working on making each of them fully editable.

Alternate HS input field
I added an alternative HS code field on the IS case input form.
It allows to manually input the HS code (just click on the pencil).
Then if you go back to the select list, it'll show you which HS category it is.

Custom categories (HS or ISIC)
As discussed, you can now create a new custom category to extend either the HS or ISIC classification.
The link is available (like the others) in the new "expert bar".

Auto-import page
It is not ready yet! but you can see how it is going to look like. I plan to finish it by next week.